# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BitBox API",
    "description": "Rust API for 360 Project notation online",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/token/reset": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Request password reset token",
        "description": "This endpoint allows users to request a password reset token which is returned in a cookie.",
        "operationId": "get_reset_token_route",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The reset token id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Responded with JWT",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
    "/token/evaluation": {
      "get": {
        "tags": [
          "Evaluation"
        ],
        "summary": "Request student token to evaluate his group",
        "description": "This endpoint allows student to get the token to evaluate his group.",
        "operationId": "get_evaluation_token_route",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The evaluation token id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Responded with JWT",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "example": "UnauthorizedError"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InternalError": {
        "default": null,
        "nullable": true
      },
...
      "UnauthorizedError": {
        "default": null,
        "nullable": true
      },
...
    },
...
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "api_key": [
        "read",
        "write"
      ]
    }
  ]
}